// src/styles/enums.ts
export enum FONT {
  p = 'var(--font-p)',
  s = 'var(--font-s)',
}

export enum FONT_SIZE {
  '2xs' = 'var(--size-font-2xs)',
  'xs' = 'var(--size-font-xs)',
  's' = 'var(--size-font-s)',
  'm' = 'var(--size-font-m)',
  'l' = 'var(--size-font-l)',
  'xl' = 'var(--size-font-xl)',
  '2xl' = 'var(--size-font-2xl)',
}

export enum NOT_FONT_SIZE {
  '6xs' = 'var(--size-not-font-6xs)',
  '5xs' = 'var(--size-not-font-5xs)',
  '4xs' = 'var(--size-not-font-4xs)',
  '3xs' = 'var(--size-not-font-3xs)',
  '2xs' = 'var(--size-not-font-2xs)',
  'xs' = 'var(--size-not-font-xs)',
  's' = 'var(--size-not-font-s)',
  'm' = 'var(--size-not-font-m)',
  'l' = 'var(--size-not-font-l)',
  'xl' = 'var(--size-not-font-xl)',
  '2xl' = 'var(--size-not-font-2xl)',
  '3xl' = 'var(--size-not-font-3xl)',
  '4xl' = 'var(--size-not-font-4xl)',
  '5xl' = 'var(--size-not-font-5xl)',
  '6xl' = 'var(--size-not-font-6xl)',
}

export enum COLOR {
  a_d2 = 'var(--color-a-d2)',
  a_d1 = 'var(--color-a-d1)',
  a = 'var(--color-a)',
  a_b1 = 'var(--color-a-b1)',
  a_b2 = 'var(--color-a-b2)',

  b_d2 = 'var(--color-b-d2)',
  b_d1 = 'var(--color-b-d1)',
  b = 'var(--color-b)',
  b_b1 = 'var(--color-b-b1)',
  b_b2 = 'var(--color-b-b2)',

  c_d2 = 'var(--color-c-d2)',
  c_d1 = 'var(--color-c-d1)',
  c = 'var(--color-c)',
  c_b1 = 'var(--color-c-b1)',
  c_b2 = 'var(--color-c-b2)',

  d_d2 = 'var(--color-d-d2)',
  d_d1 = 'var(--color-d-d1)',
  d = 'var(--color-d)',
  d_b1 = 'var(--color-d-b1)',
  d_b2 = 'var(--color-d-b2)',

  g_19 = 'var(--color-g-19)',
  g_18 = 'var(--color-g-18)',
  g_17 = 'var(--color-g-17)',
  g_16 = 'var(--color-g-16)',
  g_15 = 'var(--color-g-15)',
  g_14 = 'var(--color-g-14)',
  g_13 = 'var(--color-g-13)',
  g_12 = 'var(--color-g-12)',
  g_11 = 'var(--color-g-11)',
  g_10 = 'var(--color-g-10)',
  g_9 = 'var(--color-g-9)',
  g_8 = 'var(--color-g-8)',
  g_7 = 'var(--color-g-7)',
  g_6 = 'var(--color-g-6)',
  g_5 = 'var(--color-g-5)',
  g_4 = 'var(--color-g-4)',
  g_3 = 'var(--color-g-3)',
  g_2 = 'var(--color-g-2)',
  g_1 = 'var(--color-g-1)',
  g_0 = 'var(--color-g-0)',
}

export enum MICROINTERACTION {
  xs = 'var(--time-microinteraction-xs)',
  s = 'var(--time-microinteraction-s)',
  m = 'var(--time-microinteraction-m)',
  l = 'var(--time-microinteraction-l)',
  xl = 'var(--time-microinteraction-xl)',
}